Pull json button to own component
